More overloading for memory management
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@8345 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/support-values.hpp
@@ -85,7 +85,7 @@ namespace Gecode { namespace Int {
   inline
   SupportValues<View,A>::~SupportValues(void) {
     a.free(bits,sz);
-    a.free(rp_fst,static_cast<unsigned int>(rp_lst-rp_fst+1));
+    a.free(rp_fst,rp_lst-rp_fst+1);
   }
 
   template<class View, class A>
